// src/app/pages/operaciones/registro-de-contratos/impresiones/contrato-impresion.interface.ts
export interface ContratoImpresion {
  idContrato: number,
  numeroContrato: string | null;
  strNombreIEC: string | null;
  origenVenta: string | null;
  nombreContratante: string | null;
  apellidoPaternoContratante: string | null;
  apellidoMaternoContratante: string | null;
  numeroDocumentoContratante: string | null;
  rucContratante: string | null;
  fechaNacimientoContratante: string | null;
  estadoCivilContratante: string | null;
  parentescoContratatante: string | null;
  nacionalidadContratante: string | null;
  telefonoContratante: string | null;
  direccionContratante: string | null;
  correoContrantante: string | null,
  nombreContratante2: string | null;
  apellidoPaternoContratante2: string | null;
  apellidoMaternoContratante2: string | null;
  numeroDocumentoContratante2: string | null;
  rucContratante2: string | null;
  fechaNacimientoContratante2: string | null;
  estadoCivilContratante2: string | null;
  parentescoContratatante2: string | null;
  nacionalidadContratante2: string | null;
  telefonoContratante2: string | null;
  direccionContratante2: string | null;
  correoContrantante2: string | null,
  nombreBeneficiario: string | null;
  apellidoPaternoBeneficiario: string | null;
  apellidoMaternoBeneficiario: string | null;
  documentoBeneficiario: string | null;
  fechaNacimientoBeneficiario: string | null;
  direccionBeneficiario: string | null;
  nombreFallecido: string | null;
  apellidoPaternoFallecido: string | null;
  apellidoMaternoFallecido: string | null;
  documentoFallecido: string | null;
  fechaNacimientoFallecido: string | null;
  fechaFallecimientoFallecido: string | null;
  fechaEntierroFallecido: string | null;
  nacionalidadFallecido: string | null;
  codigoPlan: string | null;
  nombrePlan: string | null;
  tipoSepultura: string | null;
  capacidadContratada: string | null;
  capacidadTotal: string | null;
  urnas: string | null;
  servicioInhumacion: string | null;
  plataforma: string | null;
  manzana: string | null;
  numeroSepultura: string | null;
  nivel: string | null;
  codigoSepultura: string | null;
  numeroReserva: string | null;
  precioLista: number | null;
  descuento: number | null;
  inhumacion: number | null;
  foma: number | null;
  precioTotal: number | null;
  pagoInicial: number | null;
  saldo: number | null;
  numeroCuotas: number | null;
  montoCuota: number | null;
  fechaCuota: string | null;
  ultimoVencimiento: string | null;
  fechaContrato: string | null;
  comprobante: string | null;
  sistemaPago: string | null;
  documentoFinanciado: string | null,
  tipoContrato: string | null,
  tipoDeceso: string | null,
  religionFallecido: string | null
}